requirements.txt
discord.py
essentia
numpy
youtube_dl
